reset gamethread
